Use context factories instead of direct contexts
A DbContext instance can only do one thing at a time and is often advised to be used for a single unit of work and consequently thrown away.
The current system supplies a single DbContext to an entire service, which consists of multiple asynchronous methods which could possibly run in parallel.

To prevent any future issues I have taken the liberty of transitioning to a factory model, now contexts are created for every execution of a public service method.
I had to change this because I was having trouble with both the navigation header & login page trying to access UserService.GetUser() at the same time.